some db changes

+++ server/database/migrations/2024_01_27_220415_create_wallets_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->ulid('id')->primary();
             $table->foreignUlid('user_id')->constrained();
-            $table->unsignedInteger('balance')->default(0);
-            $table->unsignedInteger('total_inflow')->default(0);
-            $table->integer('total_expenditure')->default(0);
+            $table->float('balance')->default(0);
+            $table->float('total_inflow')->default(0);
+            $table->float('total_expenditure')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
