+++ server/database/migrations/2024_02_11_094119_create_internaltransactions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('internaltransactions', function (Blueprint $table) {
             $table->ulid('id')->primary();
             $table->enum('transaction_type', ['commission_on_pay_per_view', 'commission_on_subscription', 'commission_on_stream_tip', 'commission_on_tip', 'commission_on_vat']);
-            $table->unsignedInteger('amount');
+            $table->float('amount');
             $table->ulid('reference_id_to_resource');
             $table->ulid('reference_id_to_transaction');
             $table->timestamps();
